Allow error handlers to resolve error and return result (#3)
Pass ErrorHandlerState to error handlers. The handler can then call #set_as_handled with an optional result. This will mark the error as "handled" and the mediator will skip any subsequent error handlers. If the exception was thrown when handling a request, the result passed to set_as_handled will be returned to the message dispatcher.